backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
python-dotenv==1.0.1

# AI
google-generativeai==0.8.3
openai==1.47.1

# Google Calendar
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Supabase
supabase==2.7.4

# Scheduling (for cron jobs)
apscheduler==3.10.4
